Fix quantity not being treated as decimal value, Add transactions screen, Refactor fake data generations for previews

+++ app/src/main/java/com/kssidll/arrugarq/ui/screen/additem/AddItemData.kt
@@ -6,7 +6,7 @@ data class AddItemData(
     var productId: Long,
     var variantId: Optional<Long>,
     var shopId: Optional<Long>,
-    var quantity: Long,
+    var quantity: Float,
     var price: Float,
     var date: Long,
 )
